minor changes in repository and model classes just annotations and validations

+++ ai-job-mail/src/main/java/com/demo/repository/EmailRequestRepository.java
@@ -1,9 +1,8 @@
 package com.demo.repository;
 
-import com.demo.model.*;
-
+import com.demo.model.EmailRequest;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EmailRequestRepository extends JpaRepository<EmailRequest,Integer>{
-
+public interface EmailRequestRepository extends JpaRepository<EmailRequest, Long> {
+    // No changes needed — just ID type corrected to Long
 }
